Unify pre-commit markdown sync scripts and refine documentation
Unify the separate sync scripts into scripts/sync_md_examples.py with
individual pre-commit hook stanzas for README and TYPE_ALGEBRA. Apply
prose shorthand rule to omit fn:: prefixes on core vocabulary types in
README.md, and remove the temporal 'Now' from CONTRIBUTING.md. Use
C++20 trailing return type styles in README.md and its compiled example.
